requirements.txt
Django==3.2.5
django-crispy-forms==1.12.0
Pillow==8.3.1
django-cleanup==5.2.0
django-ckeditor==6.1.0
django-recaptcha==2.0.6
